Move App.js
