+++ client/src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import './css/index.css';
-import App from './components/App';
+import App from './App';
 import * as serviceWorker from './serviceWorker';
+import './css/index.css';
 
 ReactDOM.render(<App />, document.getElementById('root'));
 serviceWorker.unregister();
